refactor: Replace context.TODO() with proper context in FlightCtlManager
The use of context.TODO() is generally discouraged in production code as it should only be used when it's unclear which context to use. In this case, we have a proper context available from the StartReconcileFlightCtlResources function parameter.

This change improves code quality by: 1. Using the proper context that can be cancelled when the manager stops 2. Following Go best practices for context propagation 3. Making the code more maintainable by removing temporary solutions

Signed-off-by: xuezhaojun <zxue@redhat.com>